New formatters for binary images. Timeout for capture call.
Added formatters to deal with image formats used, correct content type
will be used instead of application/octet-stream. Using the binary
formatter implementation from Restify for both cases. Added a timeout so
that capture API call would terminate with an error if the instance
could not be acquired within a minute (this will still leave the
callback function to wait until phantom instance is provided, but it
will just be returned to the pool). Added some more gotchas to the
README file, about resource consumption in case of load testing
(probably need to add a Jmeter plan file into the repo some time).